+++ src/hooks/index.ts
@@ -22,3 +22,4 @@ export * from './context/usePoolAccountsContext';
 export * from './context/useAccountContext';
 export * from './context/useAuthContext';
 export * from './context/useNotificationsContext';
+export * from './useFeatureFlag';